Add prototype romanesco AST parser
This commit adds a prototype AnalysisStaticParser class that Parses
files with romanesco python and spark.python tasks using abstract syntax
trees. The returned parse is valid romanesco analysis JSON. This is done
by wrapping injected variables in metaclasses that derive from
RomanescoTypeConverter. In a script this looks like:

from classes import Table, Output

def main(a, b):
    c = {'fields': a['fields'], 'rows': a['rows'] + b['rows']}
    return c

if __name__ == "custom":
    c = Output(main(Table(a, format="rows"), Table(b, format="rows")),
               type="table", format="rows")

Where Table() and Output() are defined in classes and foo() is expected
to do some kind of computation on a and b. Classes inheriting from
ROmanescoTypeConverter, when executed,  currently return the variable
unchanged.  Basic infrastructure exists to allow for validation and
conversion depending on the romanesco type,  and the nature of the
injected variable. If the above script were placed into a file
"test_analysis.py" one could call the following code:

AnalysisStaticParser("test_analysis.py", "python").parse()

And it would return the following JSON:

{'inputs': [{'format': 'rows', 'name': 'a', 'type': 'table'},
            {'format': 'rows', 'name': 'b', 'type': 'table'}],
 'mode': 'python',
 'name': 'test_analysis',
 'outputs': [{'format': 'rows', 'name': 'c', 'type': 'table'}],
 'script': '...omitted...'}

To see this in action you can cd into the examples/ directory and run
'python run_example.py'